# gen/openapiv2/eigenlayer/rewards_service/v1/api.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "eigenlayer/rewards_service/v1/api.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "Rewards"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/claim-proof": {
      "post": {
        "operationId": "Rewards_GenerateClaimProof",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GenerateClaimProofResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1GenerateClaimProofRequest"
            }
          }
        ],
        "tags": [
          "Rewards"
        ]
      }
    },
    "/v1/earned-tokens": {
      "post": {
        "operationId": "Rewards_GetEarnedTokensForEarner",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetEarnedTokensForEarnerResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1GetEarnedTokensForEarnerRequest"
            }
          }
        ],
        "tags": [
          "Rewards"
        ]
      }
    }
  },
  "definitions": {
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "v1EarnerLeaf": {
      "type": "object",
      "properties": {
        "earner": {
          "type": "string"
        },
        "earnerTokenRoot": {
          "type": "string"
        }
      }
    },
    "v1GenerateClaimProofRequest": {
      "type": "object",
      "properties": {
        "earnerAddress": {
          "type": "string"
        },
        "tokens": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "v1GenerateClaimProofResponse": {
      "type": "object",
      "properties": {
        "proof": {
          "$ref": "#/definitions/v1Proof"
        }
      }
    },
    "v1GetEarnedTokensForEarnerRequest": {
      "type": "object",
      "properties": {
        "earnerAddress": {
          "type": "string"
        }
      }
    },
    "v1GetEarnedTokensForEarnerResponse": {
      "type": "object",
      "properties": {
        "tokens": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1TokenAmount"
          }
        }
      }
    },
    "v1Proof": {
      "type": "object",
      "properties": {
        "root": {
          "type": "string"
        },
        "rootIndex": {
          "type": "integer",
          "format": "int64"
        },
        "earnerIndex": {
          "type": "integer",
          "format": "int64"
        },
        "earnerTreeProof": {
          "type": "string"
        },
        "earnerLeaf": {
          "$ref": "#/definitions/v1EarnerLeaf"
        },
        "leafIndices": {
          "type": "array",
          "items": {
            "type": "integer",
            "format": "int64"
          }
        },
        "tokenTreeProofs": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "tokenLeaves": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1TokenLeaf"
          }
        }
      }
    },
    "v1TokenAmount": {
      "type": "object",
      "properties": {
        "token": {
          "type": "string"
        },
        "amount": {
          "type": "string"
        }
      }
    },
    "v1TokenLeaf": {
      "type": "object",
      "properties": {
        "token": {
          "type": "string"
        },
        "cumulativeEarnings": {
          "type": "string"
        }
      }
    }
  }
}